+++ lib/hexapdf/digital_signature/signing/default_handler.rb
@@ -211,6 +211,13 @@ module HexaPDF
         # The contact information. If used, will be set on the signature dictionary.
         attr_accessor :contact_info
 
+        # The custom signing time.
+        #
+        # The signing time is usually the time when signing actually happens. This is also what
+        # HexaPDF uses. If it is known that signing happened at a different point in time, that time
+        # can be provided using this accessor.
+        attr_accessor :signing_time
+
         # The size of the serialized signature that should be reserved.
         #
         # If this attribute is not set, an empty string will be signed using #sign to determine the
@@ -277,7 +284,7 @@ module HexaPDF
         def finalize_objects(_signature_field, signature)
           signature[:Filter] = :'Adobe.PPKLite'
           signature[:SubFilter] = (signature_type == :pades ? :'ETSI.CAdES.detached' : :'adbe.pkcs7.detached')
-          signature[:M] = Time.now
+          signature[:M] = self.signing_time ||= Time.now
           signature[:Reason] = reason if reason
           signature[:Location] = location if location
           signature[:ContactInfo] = contact_info if contact_info
@@ -312,6 +319,7 @@ module HexaPDF
                                      type: signature_type,
                                      certificate: certificate, key: key,
                                      digest_algorithm: digest_algorithm,
+                                     signing_time: signing_time,
                                      timestamp_handler: timestamp_handler,
                                      certificates: certificate_chain, &external_signing).to_der
           else